requirements.txt
# Core LLM & LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-google-genai>=1.0.0
langserve>=0.2.0

# Local embeddings via Ollama
ollama>=0.2.0

# Vector store
chromadb>=0.5.0

# Document parsing
unstructured[pdf]>=0.15.0
pdf2image
pillow
python-docx>=1.1.0

# FastAPI & server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
requests>=2.32.0

# Frontend
gradio==4.44.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
tenacity>=8.3.0
loguru>=0.7.0

# Evaluation
ragas>=0.1.9
datasets>=2.19.0

Tested with:
- Python 3.11
- Gradio 4.44.0